Move game mocks into a game mock directory
