+++ mocks/__init__.py
@@ -1,4 +1,3 @@
 from .socket import socket
 from .connection import Connection
-from .game_record import Game_Record
-from .players import Players
+from .game import *
